Reworking intents
Hi Austin. Here are my edits with some notes. I learned a lot
about my own expectations regarding speech recognition and my
own experience writing intents from reading your intents, and
I hope these suggestions might help us write a guide to help
new users with intent creation.

The first thing I noticed is that you are counting on specific
words appearing in specific orders. This really isn’t the way
speech recognition works right now. I am training a puppy, and
have been struck by the similarities between training the dog
and the computer. In both cases, the listener doesn’t really
understand what you are saying and is basically listening for
keywords like “SIT”, “COME”, “DROP IT”, “WALKIES”, etc. Using
a language model helps somewhat, but random noises will still
be inserted into the transcription.

You might think you said “Tell me the day of the week” but Naomi
might hear “TELL ME DAY OF A THE WEEK”. Matching the specific
phrase here doesn’t work. The default intent parser for Naomi uses
distance edits to choose the best match from the possible list of
intents it has learned. Naomi is more likely to get the gist of
what you are saying than to produce a perfect transcription.

For instance, when Naomi first started and I asked what time it was,
the background noise did not immediately drop off to the point where
the VAD recognized that I had stopped talking, so I repeated the
command more loudly. The resulting transcript was “ONE WILL TIME IS
IT NAOMI TIME IS” which is pretty good, and allowed Naomi to respond
appropriately, but trying to match “WHAT TIME IS IT” would fail.

You are also trying to extract keywords using specific slices,
(location = transcript[19:]) which is problematic since it is
extremely unlikely that transcript will exactly match the phrase
you are expecting. In this case, it is really best to use Keywords
and let the intent parser handle parsing them. You might need to do
a little checking and cleanup on the results, but I haven’t found
it necessary in general. So instead of “location = transcript[19:]”
I used “for location in intent['matches']['LocationKeyword']:”
(since a keyword returns a list – this is usually a list of one,
but if the user asks “What time is it in Hawaii and London?” it is
nice to allow Naomi to respond to both requests). For instance, if
I ask “What day is Halloween this year?” and Naomi hears “WHAT DAY
A HALLOWEEN THIS YEAR” it would be looking for the holiday “ALLOWEEN
THIS YEAR”.

On the subject of Keywords, it is important to build a vocabulary
for Naomi. I think this was one of the key discoveries by the
original Jasper developers. None of the open source speech
recognition systems are really open-ended, and trying to get them
to listen for every possible word in the English language really
increases the word error rate. So “'WHAT TIME IS IT IN {}” is not
a very helpful template. Give the keyword a name, build a list of
possible values, then use the intent matching function to figure
out what was requested. In this case, we can’t really pass a list
of every place name on earth to Naomi without overwhelming it, so
I just chose a short list of places I might ask about when making
phone calls. You might be able to find a list of common place names,
or search for a country name, then a region name to limit the list
of places you are searching for. This is part of the reason I think
speechhandler plugins need access to the WAV file of the request
being serviced, but that is still a work in progress. We can’t
really handle open ended requests like “SEARCH GOOGLE FOR {}”. We
should eventually be able to use some predictions to help Naomi
figure out what is being requested, but we haven’t done that yet,
and therein lies the art and fun of speech recognition right now.

You can also have more than one intent parser in a plugin, so instead
of using one intent and trying to figue out which action is being
requested within the handler, I split this out into three different
intents pointing to three different handlers - one for time,
one for dates, and one for holidays. That way, by the time the
intent gets to the handler, I already have a good idea of what is
being requested and can search the input text or matches for
keywords that tell me exactly what to do.

I hope that makes sense and can be helpful. Again, I think this
could be a germ of a tutorial on how transcriptions, intents, and
intent parsing work together, and I'd welcome help and suggestions
on how to present this information.